[FLINK-14504][rest] Rename AbstractTaskManagerHandler
The name was misleading since the handler only interacts with the ResourceManager, and just happened to only be used by TaskExecutor-related handlers.
